Move components around for ease of use
